requirements.txt
################### Python 3.11.8 ###################
notebook==7.1.2
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
jupyter==1.0.0
emoji==2.10.1
regex==2023.12.25
openai==1.14.2
python-dotenv==1.0.1
tqdm==4.66.2
transformers==4.38.2
scikit-learn==1.4.1.post1
datasets==2.18.0
torch==2.2.1
accelerate==0.28.0
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
fastapi-cors==0.0.6
setuptools==69.2.0
wheel==0.43.0
wandb==0.16.4

